requirements.txt
pandas~=2.2.0
pytest~=7.4.4
gitpython
pycap>=2.0.0
elabapi_python
PyYAML~=6.0.1
setuptools~=68.2.0
numpy~=1.26.3